==> code/01_CombineDatasets.R <==
## Merge the data from Fleishman et al and Chambers et al.

require(stringr)
require(reshape2) #load the necessary packages
require(dplyr)

#Load the original data. Note that I saved data from each excel workbook into csv's
#Underdown Data provided by Dave Board
infolder <- here::here("original_data/") #location of original data files
outfolder <- here::here("processed_data/") #location for files created after processing
db_original <- read.csv(paste0(infolder,"/BRTE_Count_DB.csv"), stringsAsFactors = FALSE) #original data as sent to me by David Baord on 11/30/16 from the Underdown work of Chambers et al

#Clean up data (eliminate dupes, eliminate extraneous columns, etc)
#remove duplicates
db_nd <- db_original[!duplicated(db_original[,2:22]),] #UPDATE 10 Feb 2017 Removal of Dupes is correct see email from D.Board 13 Dec 2016
#get igrass data cleaned up to be incidence not percentage
db.igrass <- db_original[,c(2,11,15)]
db.igrass <- db.igrass[db.igrass$Year== 2003,] #the burn was in 2002 so 2003 data is first growing season post-burn, this would have to be changed if we go with year pre-burn
colnames(db.igrass)[2] <- "igrass"
db.igrass <- db.igrass[,c(3,2)]

#get data frame in shape for merging
db <- db_nd[,c(13,14,15,2,8,16,17,4,18,21,11)]
db$tburn <- db$Year - db$firstGS #this makes tburn for the immediate post-fire measurements 0
db$mtype <- 0
colnames(db)[8] <- "burn"
db$burn <- as.numeric(as.factor(db$burn)) #converts to index where burn in 1 and control =2, need to convert control to 0
db$burn <- ifelse(db$burn == 2, 0, db$burn)
db$burn <- ifelse(db$Year < 2002, 0, db$burn) # sets all burn values to 0 for years before the burn
db$pgrass <- db$Pgcount/db$num_samples # convert pg to frequency


#Load Precip data provided by Erica Fleishman
db_precip <- read.csv(paste0(infolder,"/BRTE_precip_DB.csv")) #precipitation data from PRISM obtained and generated by Erica Fleishman
#melt precip  data
dbp <- db_precip[,2:34] #Can't have canyon
und.pc.melt <- melt(dbp, id.vars="point", value.name = "precip") #melt to long form
und.pc.melt$year <- str_extract(und.pc.melt$variable, "[0-9]+") #get year
und.pc.melt$season <- substr(und.pc.melt$variable, 13, 30) #extract season
und.pc.spring <- und.pc.melt[und.pc.melt$season == "spring",] #create seasonal table for merging
und.spring <- und.pc.spring[,c(1,4,3)]
und.pc.winter <- und.pc.melt[und.pc.melt$season == "winter",] #create seasonal table for merging
und.winter <- und.pc.winter[,c(1,4,3)]


#Load Grazing Data provided by Erica Fleishman
db_grazing <- read.csv(paste0(infolder,"/BRTE_grazing_DB.csv")) #indicates whether the transect occured in an allotment that was grazed and the proportion of years that allotment had been grazed (prop_). Data is from Erica Fleishman obtained from Michael West at USFS
dbg <- db_grazing[,2:22] #Can't have canyon

#drop prop grazing columns provided by EF (done because initial grazing data was incorrect)
db_g <- dbg[,-grep("_prop_", colnames(dbg))]
db_g$grazing_2006 <- 0 #per Jeanne Chambers Underdown was not grazed from 2002 - 2007
db_g$grazing_2007 <- 0
db_g$grazing_2005 <- 0
db_g$grazing_2004 <- 0
db_g$grazing_2003 <- 0
db_g$grazing_2002 <- 0
db_g$grazing_2001 <- 0

a <- db_g[ , order(names(db_g))] #order columns by ascending year
und.g <- a[,c(16,1:15)]
und.g <- melt(und.g, id.vars="point", value.name = "grazed")
colnames(und.g)[2] <- "year"
und.g$year <- str_extract(und.g$year,"[0-9]+")

und.graze <- und.g %>% group_by(point) %>% mutate(cumsum=cumsum(grazed))
und.graze$propgraze <- und.graze$cumsum/(as.numeric(und.graze$year) - 2000) #estimate the proportion of years grazed prior to that year

#merge igrass, precip, and grazing
dbigrass <- merge(db, db.igrass, by = "point", all.x = TRUE )
dbigrass$f <- dbigrass$igrass/dbigrass$num_samples
colnames(dbigrass)[4] <- "year"

dbigrassgraze <- merge(dbigrass, und.graze, by = c("point","year"), all.x=TRUE)


dbigrassgrazeprecip <- merge(dbigrassgraze, und.spring, by = c("point","year"), all.x=TRUE)
colnames(dbigrassgrazeprecip)[19] <- "sppr"
dbigrassgrazeprecip2 <- merge(dbigrassgrazeprecip, und.winter, by = c("point","year"), all.x=TRUE)
colnames(dbigrassgrazeprecip2)[20] <- "wtpr"

#create elevation lookup to merge missing elev values
db.elookup <- db_nd[,c("point","elevation")]
db.db <- db.elookup[complete.cases(db.elookup),]
db.d <- unique(db.db)
dbigrassgrazeprecip3 <- merge(dbigrassgrazeprecip2, db.d, by = "point", all.x = TRUE)

und.merge <- dbigrassgrazeprecip3[,c(3,4,1,2,5,6,7,21,8,16,18,15,12,13,19,20)]
colnames(und.merge) <- c("range","area","pt","yr","brte","nsamp","mtype","elev","burn","graze","pgraze","igrassf","tburn","pgrassf","sppr","wtpr")
und.merge$brte <- ifelse(is.na(und.merge$brte),0,und.merge$brte)

#Fleishman Survey Data
ef_original <- read.csv(paste0(infolder,"/BRTE_CountCovs_EF.csv"), stringsAsFactors = FALSE) #this is all data from EF, note that file structure differs from previous iterations of the analysis eliminating code that was previously used to clean data
ef_2015 <- read.csv(paste0(infolder, "/corrected_2015.csv"), stringsAsFactors = FALSE)
ef_or_2015 <- ef_original[ef_original$year == 2015,] #ef original is misssing grass springs data ef_2015 is missing grazing data

#merging corrrect ef_2015 data with grazing data
ef_or_2015_graze <- ef_or_2015[,c(4,22:23)]
graze_merge <- merge(ef_2015, ef_or_2015_graze, by = "point", all.x=TRUE)
graze_merge <- graze_merge[,-c(21:22)]
colnames(graze_merge)[21:22] <- c("grazed.1", "prop_grazed")
ef_original$grazed <- NULL #this column is empty, grazing data is in grazed.1
ef.a <- ef_original[ef_original$year != 2015,] #remove incorrect 2015 data
ef.a <- rbind(ef.a,graze_merge)

#Fleishman data
ef.a$mtype <- 1
ef.a$tburn <- ef.a$year - ef.a$firstGS #This gives one point a -1 which either signals an error or that the site wasn't burned yet
ef.a$tburn <- ifelse(ef.a$tburn < 0, NA, ef.a$tburn) #elimnate the -1 value
ef.a$pgrass <- ef.a$grasscvr/ef.a$num_samples_grass.forb #convert pgrass counts to frequency
ef.a$igrassf <- ef.a$igrass/ef.a$num_samples_grass.forb #convert initial grass cover to frequency

ef <- ef.a[,c(2:4,8,6,5,7,9,10,21,22,25,23,24,18,17)]
colnames(ef) <- c("range","area","pt","yr","brte","nsamp","mtype","elev","burn","graze","pgraze","igrassf","tburn","pgrassf","sppr","wtpr")

db <- rbind(ef,und.merge) #combined datasets
write.csv(db, file =paste0(outfolder, "/alldatamerged.csv"), row.names = FALSE)
